+++ CHANGELOG.md
@@ -6,6 +6,7 @@ manually.
 * Fixed parsing VTT subtitles when there's no subtitles in the file (by https://github.com/trms-alex).
 * Added ES translations (by https://github.com/koldo92)
 * Updated dependencies
+* Updated iOS example configuration
 
 ## 0.0.67
 * Added support for DASH adaptive stream subtitles, audio tracks, tracks (by https://github.com/adrianByv)